// src/interfaces/User.model.ts
import { InfiniteScrollModel } from '@/interfaces/InfiniteScroll.model';

export interface UserModel extends InfiniteScrollModel {
  name: string;
  username: string;
  email: string;
  address: Address;
  phone: string;
  website: string;
  company: Company;
}

export interface Company {
  name: string;
  catchPhrase: string;
  bs: string;
}

export interface Address {
  street: string;
  suite: string;
  city: string;
  zipcode: string;
  geo: {
    lat: string;
    lng: string;
  };
}